+++ Nos/Controller/PersistenceController.swift
@@ -93,6 +93,11 @@ final class PersistenceController {
         }
     }
     
+    /// Deletes all entities in this object's container.
+    func deleteAll() async throws {
+        await container.deleteAllEntities()
+    }
+    
     private static func clearCoreData(store storeURL: URL, in container: NSPersistentContainer) {
         Log.info("Dropping Core Data...")
         do {