+++ Nos/Service/CurrentUser.swift
@@ -10,8 +10,9 @@ import Dependencies
     @ObservationIgnored @Dependency(\.crashReporting) private var crashReporting
     @ObservationIgnored @Dependency(\.persistenceController) private var persistenceController
     @ObservationIgnored @Dependency(\.pushNotificationService) private var pushNotificationService
-    @ObservationIgnored @Dependency(\.relayService) private var relayService
+    @ObservationIgnored @Dependency(\.relayService) var relayService
     @ObservationIgnored @Dependency(\.keychain) private var keychain
+    @ObservationIgnored @Dependency(\.unsAPI) var unsAPI
     
     // TODO: it's time to cache this
     var keyPair: KeyPair? {
@@ -98,13 +99,13 @@ import Dependencies
     
     // TODO: this is fragile
     // Reset CurrentUser state
-    @MainActor func reset() {
+    @MainActor private func reset() {
         onboardingRelays = []
         subscriptions = []
         setUp()
     }
     
-    @MainActor func setUp() {
+    @MainActor private func setUp() {
         if let keyPair {
             do {
                 author = try Author.findOrCreate(by: keyPair.publicKeyHex, context: viewContext)
@@ -122,6 +123,8 @@ import Dependencies
                 
                 Task {
                     await subscribe()
+                    // Listen for notifications
+                    await pushNotificationService.listen(for: self)
                     refreshFriendMetadata()
                 }
             } catch {
@@ -201,8 +204,6 @@ import Dependencies
         subscriptions.append(
             await relayService.fetchEvents(matching: importantEventsFilter)
         )
-        // Listen for notifications
-        await pushNotificationService.listen(for: self)
     }
     
     private var friendMetadataTask: Task<Void, any Error>?
@@ -284,206 +285,6 @@ import Dependencies
         return followKeys.contains(key)
     }
     
-    /// Builds a dictionary to be used as content when publishing a kind 0
-    /// event.
-    private func buildMetadataJSONObject(author: Author) -> [String: String] {
-        var metaEvent = MetadataEventJSON(
-            displayName: author.displayName,
-            name: author.name,
-            nip05: author.nip05,
-            uns: author.uns,
-            about: author.about,
-            website: author.website,
-            picture: author.profilePhotoURL?.absoluteString
-        ).dictionary
-        if let rawData = author.rawMetadata {
-            // Tack on any unsupported fields back onto the dictionary before
-            // publish.
-            do {
-                let rawJson = try JSONSerialization.jsonObject(with: rawData)
-                if let rawDictionary = rawJson as? [String: AnyObject] {
-                    for key in rawDictionary.keys {
-                        guard metaEvent[key] == nil else {
-                            continue
-                        }
-                        if let rawValue = rawDictionary[key] as? String {
-                            metaEvent[key] = rawValue
-                            Log.debug("Added \(key) : \(rawValue)")
-                        }
-                    }
-                }
-            } catch {
-                Log.debug("Couldn't parse a JSON from the user raw metadata")
-                // Continue with the metaEvent object we built previously
-            }
-        }
-        return metaEvent
-    }
-
-    @MainActor func publishMetadata() async throws {
-        guard let pubKey = publicKeyHex else {
-            Log.debug("Error: no publicKeyHex")
-            throw CurrentUserError.authorNotFound
-        }
-        guard let pair = keyPair else {
-            Log.debug("Error: no keyPair")
-            throw CurrentUserError.authorNotFound
-        }
-        guard let context = viewContext else {
-            Log.debug("Error: no context")
-            throw CurrentUserError.authorNotFound
-        }
-        guard let author = try Author.find(by: pubKey, context: context) else {
-            Log.debug("Error: no author in DB")
-            throw CurrentUserError.authorNotFound
-        }
-
-        self.author = author
-        
-        let jsonObject = buildMetadataJSONObject(author: author)
-        let data = try JSONSerialization.data(withJSONObject: jsonObject)
-        let content = String(decoding: data, as: UTF8.self) 
-
-        let jsonEvent = JSONEvent(
-            pubKey: pubKey,
-            kind: .metaData,
-            tags: [],
-            content: content
-        )
-
-        do {
-            try await relayService.publishToAll(
-                event: jsonEvent,
-                signingKey: pair,
-                context: viewContext
-            )
-        } catch {
-            Log.error(error.localizedDescription)
-            throw CurrentUserError.errorWhilePublishingToRelays
-        }
-    }
-    
-    @MainActor func publishMuteList(keys: [String]) async {
-        guard let pubKey = publicKeyHex else {
-            Log.debug("Error: no pubKey")
-            return
-        }
-        
-        let jsonEvent = JSONEvent(pubKey: pubKey, kind: .mute, tags: keys.pTags, content: "")
-        
-        if let pair = keyPair {
-            do {
-                try await relayService.publishToAll(event: jsonEvent, signingKey: pair, context: viewContext)
-            } catch {
-                Log.debug("Failed to update mute list \(error.localizedDescription)")
-            }
-        }
-    }
-    
-    @MainActor func publishDelete(for identifiers: [String], reason: String = "") async {
-        guard let pubKey = publicKeyHex else {
-            Log.debug("Error: no pubKey")
-            return
-        }
-        
-        let tags = identifiers.eTags
-        let jsonEvent = JSONEvent(pubKey: pubKey, kind: .delete, tags: tags, content: reason)
-        
-        if let pair = keyPair {
-            do {
-                try await relayService.publishToAll(event: jsonEvent, signingKey: pair, context: viewContext)
-            } catch {
-                Log.debug("Failed to delete events \(error.localizedDescription)")
-            }
-        }
-    }
-    
-    @MainActor func publishContactList(tags: [[String]]) async {
-        guard let pubKey = publicKeyHex else {
-            Log.debug("Error: no pubKey")
-            return
-        }
-        
-        guard let relays = author?.relays else {
-            Log.debug("Error: No relay service")
-            return
-        }
-
-        var relayString = "{"
-        for relay in relays {
-            if let address = relay.address {
-                relayString += "\"\(address)\":{\"write\":true,\"read\":true},"
-            }
-        }
-        relayString.removeLast()
-        relayString += "}"
-        
-        let jsonEvent = JSONEvent(pubKey: pubKey, kind: .contactList, tags: tags, content: relayString)
-        
-        if let pair = keyPair {
-            do {
-                try await relayService.publishToAll(event: jsonEvent, signingKey: pair, context: viewContext)
-            } catch {
-                Log.debug("failed to update Follows \(error.localizedDescription)")
-            }
-        }
-    }
-    
-    /// Follow by public hex key
-    @MainActor func follow(author toFollow: Author) async throws {
-        guard let followKey = toFollow.hexadecimalPublicKey else {
-            Log.debug("Error: followKey is nil")
-            return
-        }
-
-        Log.debug("Following \(followKey)")
-
-        var followKeys = await Array(socialGraph.followedKeys)
-        followKeys.append(followKey)
-        
-        // Update author to add the new follow
-        if let followedAuthor = try? Author.find(by: followKey, context: viewContext), let currentUser = author {
-            let follow = try Follow.findOrCreate(
-                source: currentUser,
-                destination: followedAuthor,
-                context: viewContext
-            )
-
-            // Add to the current user's follows
-            currentUser.follows.insert(follow)
-        }
-        
-        try viewContext.save()
-        await publishContactList(tags: followKeys.pTags)
-    }
-    
-    /// Unfollow by public hex key
-    @MainActor func unfollow(author toUnfollow: Author) async throws {
-        guard let unfollowedKey = toUnfollow.hexadecimalPublicKey else {
-            Log.debug("Error: unfollowedKey is nil")
-            return
-        }
-
-        Log.debug("Unfollowing \(unfollowedKey)")
-        
-        let stillFollowingKeys = await Array(socialGraph.followedKeys)
-            .filter { $0 != unfollowedKey }
-        
-        // Update author to only follow those still following
-        if let unfollowedAuthor = try? Author.find(by: unfollowedKey, context: viewContext), let currentUser = author {
-            // Remove from the current user's follows
-            let unfollows = Follow.follows(source: currentUser, destination: unfollowedAuthor, context: viewContext)
-
-            for unfollow in unfollows {
-                // Remove current user's follows
-                currentUser.follows.remove(unfollow)
-            }
-        }
-
-        try viewContext.save()
-        await publishContactList(tags: stillFollowingKeys.pTags)
-    }
-    
     // MARK: - NSFetchedResultsControllerDelegate
     
     @MainActor func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
@@ -491,3 +292,34 @@ import Dependencies
     }
 }
 // swiftlint:enable type_body_length
+
+extension CurrentUser {
+    
+    /// Logs the user out, deletes all their locally stored data, and updates the app state.
+    /// - Parameter appController: The ``AppController`` for updating the app state.
+    func logout(appController: AppController) async {
+        await setKeyPair(nil)
+        analytics.logout()
+        crashReporting.logout()
+        unsAPI.logout()
+        appController.configureCurrentState()
+        try? await persistenceController.deleteAll()
+    }
+    
+    /// Deletes the user's account by publishing a request to vanish and deleting all their
+    /// locally stored data.
+    /// - Parameter appController: The ``AppController`` for updating the app state.
+    ///
+    /// > Warning: This is a destructive action, so be sure that it is actually what the
+    ///            user wants.
+    func deleteAccount(appController: AppController) async throws {
+        try await publishAccountDeletedMetadata()
+        try await publishRequestToVanish()
+        
+        // Note: Publishing the empty follow list must be last before logout because
+        //       it will remove the user's relays.
+        try await publishEmptyFollowList()
+        
+        await logout(appController: appController)
+    }
+}