+++ Nos/Views/Modifiers/ReportMenuModifier.swift
@@ -15,11 +15,53 @@ struct ReportMenuModifier: ViewModifier {
     @State private var confirmReport = false
     @State private var showMuteDialog = false
     @State private var confirmationDialogState: ConfirmationDialogState<UserSelection>?
-    
+    @State private var selectedFlagOption: FlagOption?
+    @State private var selectedFlagSendOption: FlagOption?
+    @State private var showFlagSuccessView = false
+
     @Environment(\.managedObjectContext) private var viewContext
-    
-    // swiftlint:disable function_body_length
+    @Dependency(\.featureFlags) private var featureFlags
+
     func body(content: Content) -> some View {
+        Group {
+            if featureFlags.isEnabled(.newModerationFlow) {
+                newModerationFlow(content: content)
+            } else {
+                oldModerationFlow(content: content)
+            }
+        }
+    }
+
+    /// Displays the moderation flow based on the reported object type. The old flow is still displayed for the author.
+    @ViewBuilder
+    private func newModerationFlow(content: Content) -> some View {
+        switch reportedObject {
+        case .note:
+            content
+                .sheet(isPresented: $isPresented) {
+                    NavigationStack {
+                        ContentFlagView(
+                            selectedFlagOptionCategory: $selectedFlagOption,
+                            selectedSendOptionCategory: $selectedFlagSendOption,
+                            showSuccessView: $showFlagSuccessView, 
+                            flagTarget: reportedObject,
+                            sendAction: {
+                                if let selectCategory = selectedFlagOption?.category {
+                                    publishReportForNewModerationFlow(selectCategory)
+                                    showFlagSuccessView = true
+                                }
+                            }
+                        )
+                    }
+                }
+        case .author:
+            oldModerationFlow(content: content)
+        }
+    }
+
+    // swiftlint:disable function_body_length
+    @ViewBuilder
+    func oldModerationFlow(content: Content) -> some View {
         content
         // ReportCategory menu
             .confirmationDialog(unwrapping: $confirmationDialogState, action: processUserSelection)
@@ -85,7 +127,7 @@ struct ReportMenuModifier: ViewModifier {
             }
     }
     // swiftlint:enable function_body_length
-    
+
     func processUserSelection(_ userSelection: UserSelection?) {
         self.userSelection = userSelection
         
@@ -173,7 +215,7 @@ struct ReportMenuModifier: ViewModifier {
     func topLevelButtons() -> [ButtonState<UserSelection>] {
         switch reportedObject {
         case .note:
-            ReportCategoryType.noteCategories.map { category in
+            ReportCategory.noteCategories.map { category in
                 let userSelection = UserSelection.noteCategorySelected(category)
                 
                 return ButtonState(action: .send(userSelection)) {
@@ -181,7 +223,7 @@ struct ReportMenuModifier: ViewModifier {
                 }
             }
         case .author:
-            ReportCategoryType.authorCategories.map { category in
+            ReportCategory.authorCategories.map { category in
                 let userSelection = UserSelection.authorCategorySelected(category)
                 
                 return ButtonState(action: .send(userSelection)) {
@@ -190,8 +232,38 @@ struct ReportMenuModifier: ViewModifier {
             }
         }
     }
-    
-    /// Publishes a report based on user input
+
+    /// Publishes a report based on the categories the user selected for the new moderation flow.
+    private func publishReportForNewModerationFlow(_ selectedCategory: FlagCategory) {
+        if case .privacy(let privacyCategory) = selectedFlagSendOption?.category, privacyCategory == .sendToNos {
+            sendToNosForNewModerationFlow(selectedCategory)
+        } else {
+            flagPubliclyForNewModerationFlow(selectedCategory)
+        }
+    }
+
+    private func sendToNosForNewModerationFlow(_ selectedCategory: FlagCategory) {
+        if case .report(let reportCategory) = selectedCategory {
+            // Call the publisher with the extracted ReportCategory
+            ReportPublisher().publishPrivateReport(
+                for: reportedObject,
+                category: reportCategory,
+                context: viewContext
+            )
+        }
+    }
+
+    private func flagPubliclyForNewModerationFlow(_ selectedCategory: FlagCategory) {
+        if case .report(let reportCategory) = selectedCategory {
+            ReportPublisher().publishPublicReport(
+                for: reportedObject,
+                category: reportCategory,
+                context: viewContext
+            )
+        }
+    }
+
+    /// Publishes a report based on user input for the old moderation flow.
     func publishReport(_ userSelection: UserSelection?) {
         switch userSelection {
         case .sendToNos(let selectedCategory):