+++ Nos/Views/Settings/SettingsView.swift
@@ -1,12 +1,12 @@
 import SwiftUI
 import Dependencies
+import Logger
 import SwiftUINavigation
 
 let showReportWarningsKey = "com.verse.nos.settings.showReportWarnings"
 let showOutOfNetworkWarningKey = "com.verse.nos.settings.showOutOfNetworkWarning"
     
 struct SettingsView: View {
-    @Dependency(\.unsAPI) var unsAPI
     @Dependency(\.analytics) private var analytics
     @Dependency(\.crashReporting) private var crashReporting
     @Dependency(\.persistenceController) private var persistenceController
@@ -32,6 +32,7 @@ struct SettingsView: View {
 
     fileprivate enum AlertAction {
         case logout
+        case deleteAccount
     }
 
     fileprivate enum CopyButtonState {
@@ -217,6 +218,32 @@ struct SettingsView: View {
                     .padding(.vertical, 15)
             }
             .listRowGradientBackground()
+            
+            #if STAGING || DEBUG
+            if isDeleteAccountEnabled.wrappedValue {
+                ActionButton(
+                    title: .localizable.deleteMyAccount,
+                    font: .clarityBold(.title3),
+                    padding: EdgeInsets(top: 16, leading: 0, bottom: 16, trailing: 0),
+                    depthEffectColor: .actionSecondaryDepthEffect,
+                    backgroundGradient: .verticalAccentSecondary,
+                    shouldFillHorizontalSpace: true
+                ) {
+                    alert = AlertState(
+                        title: { TextState(String(localized: .localizable.deleteAccount)) },
+                        actions: {
+                            ButtonState(role: .destructive, action: .send(.deleteAccount)) {
+                                TextState(String(localized: .localizable.delete))
+                            }
+                        },
+                        message: { TextState(String(localized: .localizable.deleteAccountDescription)) }
+                    )
+                }
+                .clipShape(Capsule())
+                .listRowBackground(Color.clear)
+                .listRowInsets(EdgeInsets())
+            }
+        #endif
         }
         .scrollContentBackground(.hidden)
         .background(Color.appBg)
@@ -235,16 +262,14 @@ struct SettingsView: View {
     fileprivate func alertButtonTapped(_ action: AlertAction) async {
         switch action {
         case .logout:
-            await logout()
+            await currentUser.logout(appController: appController)
+        case .deleteAccount:
+            do {
+                try await currentUser.deleteAccount(appController: appController)
+            } catch {
+                Log.error(error)
+            }
         }
-    }
-    
-    func logout() async {
-        await currentUser.setKeyPair(nil)
-        analytics.logout()
-        crashReporting.logout()
-        unsAPI.logout()
-        appController.configureCurrentState() 
     }
 }
 
@@ -276,6 +301,19 @@ extension SettingsView {
     private var newModerationFlowToggle: some View {
         NosToggle(isOn: isNewModerationFlowEnabled, labelText: .localizable.enableNewModerationFlow)
     }
+    
+    /// Whether account deletion is enabled.
+    private var isDeleteAccountEnabled: Binding<Bool> {
+        Binding<Bool>(
+            get: { featureFlags.isEnabled(.deleteAccount) },
+            set: { featureFlags.setFeature(.deleteAccount, enabled: $0) }
+        )
+    }
+    
+    /// A toggle for account deletion that allows the user to turn the feature on or off.
+    private var deleteAccountToggle: some View {
+        NosToggle(isOn: isDeleteAccountEnabled, labelText: .localizable.enableAccountDeletion)
+    }
 }
 #endif
 
@@ -283,8 +321,11 @@ extension SettingsView {
 extension SettingsView {
     /// Controls that will appear when the app is built for STAGING.
     @MainActor private var stagingControls: some View {
-        newMediaFeatureToggle
-        newModerationFlowToggle
+        Group {
+            newMediaFeatureToggle
+            newModerationFlowToggle
+            deleteAccountToggle
+        }
     }
 }
 #endif
@@ -296,6 +337,7 @@ extension SettingsView {
         Group {
             newMediaFeatureToggle
             newModerationFlowToggle
+            deleteAccountToggle
             Text(.localizable.sampleDataInstructions)
                 .foregroundColor(.primaryTxt)
             Button(String(localized: .localizable.loadSampleData)) {