+++ CHANGELOG.md
@@ -10,6 +10,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2024-10-01Z
+
+### Release Notes
+- Disable autocorrect on NoteComposer when running on macOS. [#1460](https://github.com/planetary-social/nos/issues/1460)
+
+### Internal Changes
+- Reduced spammy "Failed to parse Follow" log messages
+- Show WebP images without link preview overlay. Currently behind the “Enable new media display” feature flag. [#1488](https://github.com/planetary-social/nos/issues/1488)
+- Show GIF overlay for animated WebP images. Currently behind the “Enable new media display” feature flag. [#1488](https://github.com/planetary-social/nos/issues/1488)
+- Fix the Crowdin GitHub integration by using the official GitHub action. [#1520](https://github.com/planetary-social/nos/issues/1520)
+- Temporarily remove Martin from the list of CODEOWNERS.
+
+## [0.1.27] - 2024-09-25Z
+
 ### Release Notes
 - Added the option to preview a note before posting it. [#1399](https://github.com/planetary-social/nos/issues/1399)
 - Fixed side menu accessibility issues. [#1444](https://github.com/planetary-social/nos/issues/1444)
@@ -22,6 +36,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added routing to profile when tapping on follow notification. [#1447](https://github.com/planetary-social/nos/issues/1447)
 - Localized follows notifications. [#1446](https://github.com/planetary-social/nos/issues/1446)
 - Fixed alert when uploading big files suggesting users pay for nostr.build. [#1321](https://github.com/planetary-social/nos/issues/1321)
+- Fixed issue where push notifications were not re-registered after account change. [#1501](https://github.com/planetary-social/nos/issues/1501)
+- Added support for NIP-62 Request to Vanish events. [#80](https://github.com/planetary-social/nos/issues/80)
+- Added Delete Account UI. [#80](https://github.com/planetary-social/nos/issues/80)
+- Fixed issue where search results weren't sorted properly. [#1485](https://github.com/planetary-social/nos/issues/1485)
+- Delete all user data when logging out. [#1534](https://github.com/planetary-social/nos/issues/1534)
+- Publish empty metadata event and empty contact list on delete account. [#1530](https://github.com/planetary-social/nos/issues/1530)
 
 ### Internal Changes
 - Use NIP-92 media metadata to display media in the proper orientation. Currently behind the “Enable new media display” feature flag. [#1172](https://github.com/planetary-social/nos/issues/1172)
@@ -30,7 +50,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Minor refactor of Event+CoreDataClass. [#1443](https://github.com/planetary-social/nos/issues/1443)
 - Refactored feature flag and added a feature flag toggle for “Enable new moderation flow” to Staging builds. [#1496](https://github.com/planetary-social/nos/issues/1496)
 - Refactored list row gradient background.
-- Reduced spammy "Failed to parse Follow" log messages
+- Added SwiftSoup to parse Open Graph metadata. [#1165](https://github.com/planetary-social/nos/issues/1165)
+- Parse Open Graph metadata whenever an event contains a URL, doesn’t have `imeta` tags, and the URL points to an HTML document. [#1425](https://github.com/planetary-social/nos/issues/1425)
+- Added a new flow to flag notes. Currently behind the “Enable new moderation flow” feature flag. [#1489](https://github.com/planetary-social/nos/issues/1489)
 
 ## [0.1.26] - 2024-09-09Z